backend/requirements.txt
alembic==1.16.5
aniso8601==10.0.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytz==2024.2
setuptools==70.3.0
six==1.17.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.15.0
Werkzeug==3.1.3
psycopg2-binary
python-dotenv
